+++ src/main/java/de/ellpeck/actuallyadditions/mod/inventory/ContainerEnergizer.java
@@ -10,6 +10,7 @@
 
 package de.ellpeck.actuallyadditions.mod.inventory;
 
+import com.mojang.datafixers.util.Pair;
 import de.ellpeck.actuallyadditions.mod.inventory.slot.ArmorSlot;
 import de.ellpeck.actuallyadditions.mod.inventory.slot.SlotItemHandlerUnconditioned;
 import de.ellpeck.actuallyadditions.mod.inventory.slot.SlotOutput;
@@ -25,6 +26,7 @@ import net.minecraft.world.inventory.Slot;
 import net.minecraft.world.item.ItemStack;
 import net.neoforged.neoforge.capabilities.Capabilities;
 
+import java.util.Map;
 import java.util.Objects;
 
 public class ContainerEnergizer extends AbstractContainerMenu {
@@ -59,9 +61,22 @@ public class ContainerEnergizer extends AbstractContainerMenu {
 
         for (int k = 0; k < 4; ++k) {
             EquipmentSlot slot = VALID_EQUIPMENT_SLOTS[k];
-            ResourceLocation resourcelocation = InventoryMenu.TEXTURE_EMPTY_SLOTS.get(slot);
-            this.addSlot(new ArmorSlot(inventory, slot, 36 + 3 - k, 102, 19 + k * 18, resourcelocation));
+            ResourceLocation resourcelocation = InventoryMenu.TEXTURE_EMPTY_SLOTS.getOrDefault(slot, InventoryMenu.EMPTY_ARMOR_SLOT_SHIELD);
+            this.addSlot(new ArmorSlot(inventory, slot, 36 + 3 - k, 102, 19 + k * 18, resourcelocation) {});
         }
+
+        this.addSlot(new Slot(inventory, 40, 120, 19 + 3 * 18) {
+            @Override
+            public void setByPlayer(ItemStack newStack, ItemStack oldStack) {
+                inventory.player.onEquipItem(EquipmentSlot.OFFHAND, oldStack, newStack);
+                super.setByPlayer(newStack, oldStack);
+            }
+
+            @Override
+            public Pair<ResourceLocation, ResourceLocation> getNoItemIcon() {
+                return Pair.of(InventoryMenu.BLOCK_ATLAS, InventoryMenu.EMPTY_ARMOR_SLOT_SHIELD);
+            }
+        });
     }
 
     @Override