+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added the ability for fluid collectors to collect water and lava from filled cauldrons.
 * Explicitly made the Drills unbreakable.
 * Fixed worldgenned AA plants breaking on block update.
+* Add off-hand slot to Energizer/Enervator
 
 # 1.3.10+mc1.21.1
 * Fixed Fluid placer not being harvestable.